add option to process buffer key in delayed processing
